+++ tests/StackExchange.Redis.Tests/ClusterShardedTests.cs
@@ -1,4 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
@@ -173,5 +176,175 @@ public class ClusterShardedTests(ITestOutputHelper output) : TestBase(output)
         {
             Log("Slot already migrated.");
         }
+    }
+
+    [Theory]
+    [InlineData(true)]
+    [InlineData(false)]
+    public async Task SubscribeToWrongServerAsync(bool sharded)
+    {
+        // the purpose of this test is to simulate subscribing while a node move is happening, i.e. we send
+        // the SSUBSCRIBE to the wrong server, get a -MOVED, and redirect; in particular: do we end up *knowing*
+        // where we actually subscribed to?
+        //
+        // note: to check our thinking, we also do this for regular non-sharded channels too; the point here
+        // being that this should behave *differently*, since there will be no -MOVED
+        var name = $"{Me()}:{Guid.NewGuid()}";
+        var channel = sharded ? RedisChannel.Sharded(name) : RedisChannel.Literal(name).WithKeyRouting();
+        await using var conn = Create(require: RedisFeatures.v7_0_0_rc1);
+
+        var asKey = (RedisKey)(byte[])channel!;
+        Assert.False(asKey.IsEmpty);
+        var shouldBeServer = conn.GetServer(asKey); // this is where it *should* go
+
+        // now intentionally choose *a different* server
+        var server = conn.GetServers().First(s => !Equals(s.EndPoint, shouldBeServer.EndPoint));
+        Log($"Should be {Format.ToString(shouldBeServer.EndPoint)}; routing via {Format.ToString(server.EndPoint)}");
+
+        var subscriber = Assert.IsType<RedisSubscriber>(conn.GetSubscriber());
+        var serverEndpoint = conn.GetServerEndPoint(server.EndPoint);
+        Assert.Equal(server.EndPoint, serverEndpoint.EndPoint);
+        var queue = await subscriber.SubscribeAsync(channel, server: serverEndpoint);
+        await Task.Delay(50);
+        var actual = subscriber.SubscribedEndpoint(channel);
+
+        if (sharded)
+        {
+            // we should end up at the correct node, following the -MOVED
+            Assert.Equal(shouldBeServer.EndPoint, actual);
+        }
+        else
+        {
+            // we should end up where we *actually sent the message* - there is no -MOVED
+            Assert.Equal(serverEndpoint.EndPoint, actual);
+        }
+
+        Log("Unsubscribing...");
+        await queue.UnsubscribeAsync();
+        Log("Unsubscribed.");
+    }
+
+    [Fact]
+    public async Task KeepSubscribedThroughSlotMigrationAsync()
+    {
+        await using var conn = Create(require: RedisFeatures.v7_0_0_rc1, allowAdmin: true);
+        var name = $"{Me()}:{Guid.NewGuid()}";
+        var channel = RedisChannel.Sharded(name);
+        var subscriber = conn.GetSubscriber();
+        var queue = await subscriber.SubscribeAsync(channel);
+        await Task.Delay(50);
+        var actual = subscriber.SubscribedEndpoint(channel);
+        Assert.NotNull(actual);
+
+        var asKey = (RedisKey)(byte[])channel!;
+        Assert.False(asKey.IsEmpty);
+        var slot = conn.GetHashSlot(asKey);
+        var viaMap = conn.ServerSelectionStrategy.Select(slot, RedisCommand.SSUBSCRIBE, CommandFlags.None, allowDisconnected: false);
+
+        Log($"Slot {slot}, subscribed to {Format.ToString(actual)} (mapped to {Format.ToString(viaMap?.EndPoint)})");
+        Assert.NotNull(viaMap);
+        Assert.Equal(actual, viaMap.EndPoint);
+
+        var oldServer = conn.GetServer(asKey); // this is where it *should* go
+
+        using (var timeout = new CancellationTokenSource(TimeSpan.FromSeconds(5)))
+        {
+            // now publish... we *expect* things to have sorted themselves out
+            var msg = Guid.NewGuid().ToString();
+            var count = await subscriber.PublishAsync(channel, msg);
+            Assert.Equal(1, count);
+
+            Log("Waiting for message on original subscription...");
+            var received = await queue.ReadAsync(timeout.Token);
+            Log($"Message received: {received.Message}");
+            Assert.Equal(msg, (string)received.Message!);
+        }
+
+        // now intentionally choose *a different* server
+        var newServer = conn.GetServers().First(s => !Equals(s.EndPoint, oldServer.EndPoint));
+
+        var nodes = await newServer.ClusterNodesAsync();
+        Assert.NotNull(nodes);
+        var fromNode = nodes[oldServer.EndPoint]?.NodeId;
+        var toNode = nodes[newServer.EndPoint]?.NodeId;
+        Assert.NotNull(fromNode);
+        Assert.NotNull(toNode);
+        Assert.Equal(oldServer.EndPoint, nodes.GetBySlot(slot)?.EndPoint);
+
+        var ep = subscriber.SubscribedEndpoint(channel);
+        Log($"Endpoint before migration: {Format.ToString(ep)}");
+        Log($"Migrating slot {slot} to {Format.ToString(newServer.EndPoint)}; node {fromNode} -> {toNode}...");
+
+        // see https://redis.io/docs/latest/commands/cluster-setslot/#redis-cluster-live-resharding-explained
+        WriteLog("IMPORTING", await newServer.ExecuteAsync("CLUSTER", "SETSLOT", slot, "IMPORTING", fromNode));
+        WriteLog("MIGRATING", await oldServer.ExecuteAsync("CLUSTER", "SETSLOT", slot, "MIGRATING", toNode));
+
+        while (true)
+        {
+            var keys = (await oldServer.ExecuteAsync("CLUSTER", "GETKEYSINSLOT", slot, 100)).AsRedisKeyArray()!;
+            Log($"Migrating {keys.Length} keys...");
+            if (keys.Length == 0) break;
+            foreach (var key in keys)
+            {
+                await conn.GetDatabase().KeyMigrateAsync(key, newServer.EndPoint, migrateOptions: MigrateOptions.None);
+            }
+        }
+
+        WriteLog("NODE (old)", await newServer.ExecuteAsync("CLUSTER", "SETSLOT", slot, "NODE", toNode));
+        WriteLog("NODE (new)", await oldServer.ExecuteAsync("CLUSTER", "SETSLOT", slot, "NODE", toNode));
+
+        void WriteLog(string caption, RedisResult result)
+        {
+            if (result.IsNull)
+            {
+                Log($"{caption}: null");
+            }
+            else if (result.Length >= 0)
+            {
+                var arr = result.AsRedisValueArray()!;
+                Log($"{caption}: {arr.Length} items");
+                foreach (var item in arr)
+                {
+                    Log($"  {item}");
+                }
+            }
+            else
+            {
+                Log($"{caption}: {result}");
+            }
+        }
+
+        Log("Migration initiated; checking node state...");
+        await Task.Delay(100);
+        ep = subscriber.SubscribedEndpoint(channel);
+        Log($"Endpoint after migration: {Format.ToString(ep)}");
+        Assert.True(
+            ep is null || ep == newServer.EndPoint,
+            "Target server after migration should be null or the new server");
+
+        nodes = await newServer.ClusterNodesAsync();
+        Assert.NotNull(nodes);
+        Assert.Equal(newServer.EndPoint, nodes.GetBySlot(slot)?.EndPoint);
+        await conn.ConfigureAsync();
+        Assert.Equal(newServer, conn.GetServer(asKey));
+
+        using (var timeout = new CancellationTokenSource(TimeSpan.FromSeconds(5)))
+        {
+            // now publish... we *expect* things to have sorted themselves out
+            var msg = Guid.NewGuid().ToString();
+            var count = await subscriber.PublishAsync(channel, msg);
+            Assert.Equal(1, count);
+
+            Log("Waiting for message on moved subscription...");
+            var received = await queue.ReadAsync(timeout.Token);
+            Log($"Message received: {received.Message}");
+            Assert.Equal(msg, (string)received.Message!);
+            ep = subscriber.SubscribedEndpoint(channel);
+            Log($"Endpoint after receiving message: {Format.ToString(ep)}");
+        }
+
+        Log("Unsubscribing...");
+        await queue.UnsubscribeAsync();
+        Log("Unsubscribed.");
     }
 }